move job generator into own file
