+++ waspc/waspc.cabal
@@ -208,6 +208,7 @@ library
       Wasp.Generator.ServerGenerator.Common
       Wasp.Generator.ServerGenerator.ConfigG
       Wasp.Generator.ServerGenerator.ExternalCodeGenerator
+      Wasp.Generator.ServerGenerator.JobGenerator
       Wasp.Generator.ServerGenerator.OperationsG
       Wasp.Generator.ServerGenerator.OperationsRoutesG
       Wasp.Generator.ServerGenerator.Setup
